requirements.txt
ib_insync
influxdb-client
python-dotenv
flask
requests